commit3

+++ src/presentation/user/controller.ts
@@ -3,6 +3,7 @@ import { Response, Request } from "express";
 import { UserService } from "../services/user.service";
 import { UpdateUserDTO } from "../../domain/dtos/update-user.dto";
 import { error } from "console";
+import { protectAcountOuner } from "../../config/validate-ouner";
 
 export class UserController {
   constructor(private readonly userService: UserService) {}
@@ -40,6 +41,14 @@ export class UserController {
   };
   updateUser = (req: Request, res: Response) => {
     const { id } = req.params;
+    const sessionUserId = req.body.sessionUser.id;
+
+    if (!protectAcountOuner(id, sessionUserId)) {
+      return res
+        .status(401)
+        .json({ message: "you are not the owner of this account" });
+    }
+
     const [error, updateUserDto] = UpdateUserDTO.create(req.body);
 
     if (error) return res.status(422).json({ message: error });
@@ -51,6 +60,14 @@ export class UserController {
   };
   deleteUser = (req: Request, res: Response) => {
     const { id } = req.params;
+    const sessionUserId = req.body.sessionUser.id;
+
+    if (!protectAcountOuner(id, sessionUserId)) {
+      return res
+        .status(401)
+        .json({ message: "you are not the owner of this account" });
+    }
+
     this.userService
       .deleteUser(id)
       .then((data) => res.status(204).json(data))
@@ -60,8 +77,8 @@ export class UserController {
     const { email, password } = req.body;
 
     this.userService
-    .login(email, password)
-    .then((data) => res.status(200).json(data))
-    .catch((error:any) => this.handleError(error,res));
+      .login(email, password)
+      .then((data) => res.status(200).json(data))
+      .catch((error: any) => this.handleError(error, res));
   };
 }
